fix: fix feedback-router.ts

+++ src/routes/feedback-router.ts
@@ -11,17 +11,18 @@ export const feedbackRouter = () => {
 
     router.put('/:commentId', validationComment, authMiddlewareBearer, inputValidationMiddleware, async (req: Request, res: Response) => {
         const foundComment = await feedbackService.findCommentById(req.params.commentId)
+        if (!foundComment) return res.sendStatus(HTTP_STATUS.NOT_FOUND_404)
         if (req.user!.id !== foundComment!.commentatorInfo.userId) return res.sendStatus(HTTP_STATUS.FORBIDDEN_403)
 
         const updatedComment = await feedbackService.updateCommentById(req.params.commentId, req.body.content)
         if (!updatedComment.data) return res.sendStatus(convertResultErrorCodeToHttp(updatedComment.code))
-
         res.status(HTTP_STATUS.NO_CONTENT_204).json(updatedComment)
     })
 
     router.delete('/:commentId', authMiddlewareBearer, async (req: Request, res: Response) => {
         const foundComment = await feedbackService.findCommentById(req.params.commentId)
         if (!foundComment) res.sendStatus(HTTP_STATUS.NOT_FOUND_404)
+        if (req.user!.id !== foundComment!.commentatorInfo.userId) return res.sendStatus(HTTP_STATUS.FORBIDDEN_403)
 
         await feedbackService.deleteCommentById(req.params.commentId)
         res.sendStatus(HTTP_STATUS.NO_CONTENT_204)
